get online replay viewing for 0.6 replays sort of working (not crashing)
it should be relatively easy to get pre 0.6 replays working once this is finished

since online highscores send timestamps instead of noterows we match nerv noterows with replay data offsets before feeding everything into setupscore

since online replay highscores aren't actually in the profile, they aren't contained within scoreman, this adds a hacky way to temporarily store and retrieve online highscores that we don't want written to the profile, this involve dumb bool setting in stagestats, screenselect, and nothing actually clears the temp replay highscore, it just gets overwritten

im 99% sure the changes in the evalscreen and the scoreman gettempreplayscore function are no longer needed and can be purged but i dont want to test that and we may need them if we change how things work

there is a less than faithful conversion with online score playbacks, the change in player.cpp is because we're getting occasional nonsensical tns enums (-123123), and this results in replays failing out which is what the gameplay check is for

+++ src/PlayerAI.cpp
@@ -202,7 +202,7 @@ PlayerAI::SetScoreData(HighScore* pHighScore)
 
 	// Generate vectors made of pregenerated HoldReplayResults referenced by the
 	// song row in a map
-	for (int i = 0; i < (int)replayHoldVector.size(); i++) {
+	for (size_t i = 0; i < replayHoldVector.size(); i++) {
 		// Create or append to the vector
 		if (m_ReplayHoldMap.count(replayHoldVector[i].row) != 0) {
 			m_ReplayHoldMap[replayHoldVector[i].row].push_back(
